[WIP] Be adjusting geosearch state
