+++ package.json
@@ -20,6 +20,7 @@
     "react-redux": "^7.1.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "^3.2.0",
+    "react-tabs": "^3.0.0",
     "react-test-renderer": "^16.10.1",
     "redux": "^4.0.4",
     "styled-components": "^4.4.1"
@@ -67,6 +68,7 @@
     "@types/react-dom": "16.9.0",
     "@types/react-redux": "^7.1.2",
     "@types/react-router-dom": "^4.3.5",
+    "@types/react-tabs": "^2.3.1",
     "@types/styled-components": "^4.4.0",
     "@types/uuid": "^3.4.6",
     "@types/wordpress__i18n": "^3.4.0",
